# swagger/api-swagger.yml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
basePath: /
tags:
  - name: mail-controller
    description: Mail Controller
paths:
  /v1/group/mail/publish:
    post:
      tags:
        - mail-controller
      summary: publishMail
      operationId: publishMailUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: email
          description: email
          required: true
          schema:
            $ref: '#/definitions/Email'
      responses:
        '200':
          description: OK
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /v1/group/mail/sent:
    post:
      tags:
        - mail-controller
      summary: sendMail
      operationId: sendMailUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: email
          description: email
          required: true
          schema:
            $ref: '#/definitions/Email'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Response'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      deprecated: false
definitions:
  Email:
    type: object
    properties:
      attachmentUriList:
        type: array
        items:
          type: string
      content:
        type: string
      from:
        type: string
      isHtmlContent:
        type: boolean
      subject:
        type: string
      to:
        type: string
    title: Email
  Response:
    type: object
    properties:
      errorDescription:
        type: string
      success:
        type: boolean
    title: Response